requirements.txt
streamlit
pdfplumber
faiss-cpu
sentence-transformers
python-dotenv
requests
supabase
